+++ src/RestApi/Device/DeviceManagement.ts
@@ -22,6 +22,7 @@ import DeviceResolutionManagement from "./Resolution/DeviceResolutionManagement"
 import DeviceTimerManagement from "./Timer/DeviceTimerManagement";
 import DeviceVerificationManagement from "./Verification/DeviceVerificationManagement";
 import DeviceAppletTestManagement from "./AppletTest/DeviceAppletTestManagement";
+import DeviceTelemetryManagement from "./Telemetry/DeviceTelemetryManagement";
 
 export const RESOURCE: string = 'device';
 
@@ -46,6 +47,7 @@ export default class DeviceManagement {
 	public timer: DeviceTimerManagement;
 	public verification: DeviceVerificationManagement;
 	public appletTest: DeviceAppletTestManagement;
+	public telemetry: DeviceTelemetryManagement;
 
 	constructor(private accountOptions: IOptions, private organizationOptions: IOptions) {
 		this.appVersion = new DeviceAppVersionManagement(organizationOptions);
@@ -67,6 +69,7 @@ export default class DeviceManagement {
 		this.timer = new DeviceTimerManagement(organizationOptions);
 		this.verification = new DeviceVerificationManagement(organizationOptions);
 		this.appletTest = new DeviceAppletTestManagement(organizationOptions);
+		this.telemetry = new DeviceTelemetryManagement(organizationOptions);
 	}
 
 	public async list(filter: IDeviceFilter = {}): Promise<IDevice[]> {