+++ src/RestApi/Device/Telemetry/DeviceTelemetry.ts
@@ -1,0 +1,11 @@
+import { fillDataToEntity } from "../../mapper";
+import IDeviceTelemetry from './IDeviceTelemetry';
+
+export default class DeviceTelemetry implements IDeviceTelemetry {
+	public readonly updatedAt: IDeviceTelemetry['updatedAt'];
+	public readonly data: IDeviceTelemetry['data'];
+
+	constructor(data: IDeviceTelemetry) {
+		fillDataToEntity(this, data);
+	}
+}