+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Policy management
+- Latest telemetry data for given device & type
 
 ## [1.7.1] - 2021-12-17
 ### Fixed